# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AUM Packaging Customer Order Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders/{orderId}:
    get:
      operationId: getOrder
      tags: [orders]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
...
  /orders/{orderId}/quotation:
    post:
      operationId: createQuotation
      tags: [quotations]
      summary: Sales sends a quotation (or revision) for an order
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuotationRequest"
      responses:
        "200":
          description: Quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
  /orders/{orderId}/quotation/respond:
    post:
      operationId: respondToQuotation
      tags: [quotations]
      summary: Customer accepts, rejects or requests revision
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuotationResponseRequest"
      responses:
        "200":
          description: Updated quotation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quotation"
...
  /orders/{orderId}/payments:
    get:
      operationId: listOrderPayments
      tags: [payments]
      parameters:
        - $ref: "#/components/parameters/OrderId"
      responses:
        "200":
          description: Payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Accounts records an invoice / payment update
      parameters:
        - $ref: "#/components/parameters/OrderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{paymentId}/proofs:
    post:
      operationId: submitPaymentProof
      tags: [payments]
      summary: Customer uploads payment proof
      parameters:
        - $ref: "#/components/parameters/PaymentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPaymentProofRequest"
      responses:
        "200":
          description: Proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentProof"
  /payment-proofs/{proofId}/review:
    post:
      operationId: reviewPaymentProof
      tags: [payments]
      summary: Accounts approves or rejects payment proof
      parameters:
        - $ref: "#/components/parameters/ProofId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewPaymentProofRequest"
      responses:
        "200":
          description: Reviewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentProof"
...
components:
  parameters:
...
    OrderId:
      in: path
      name: orderId
      required: true
      schema:
        type: string
    PaymentId:
      in: path
      name: paymentId
      required: true
      schema:
        type: string
    ProofId:
      in: path
      name: proofId
      required: true
      schema:
        type: string
...
  schemas:
...
    OrderStatus:
      type: string
      enum:
        - requirement_received
        - quotation_sent
        - quote_revision_requested
        - quote_accepted
        - order_confirmed
        - in_production
        - ready_for_dispatch
        - partially_dispatched
        - fully_dispatched
        - invoice_generated
        - payment_pending
        - payment_received
        - order_closed
    QuotationStatus:
      type: string
      enum: [sent, accepted, rejected, revision_requested]
    PaymentStatus:
      type: string
      enum: [pending, part_paid, paid, overdue]
    PaymentProofStatus:
      type: string
      enum: [submitted, approved, rejected]
    ProductCategory:
      type: string
      enum: [preform, bottle, cap, closure]
    DispatchDocType:
      type: string
      enum: [invoice, eway_bill, lr_copy, delivery_challan, packing_list]
...
    Order:
      type: object
      required:
        - id
        - orderNumber
        - customerId
        - customerName
        - productId
        - productName
        - quantity
        - status
        - createdAt
      properties:
        id:
          type: string
        orderNumber:
          type: string
        customerId:
          type: string
        customerName:
          type: string
        productId:
          type: string
        productName:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        neckType:
          type: string
          nullable: true
        gramWeight:
          type: number
          nullable: true
        color:
          type: string
          nullable: true
        quantity:
          type: integer
        requiredDeliveryDate:
          type: string
          format: date
          nullable: true
        deliveryLocation:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        attachmentUrl:
          type: string
          nullable: true
        attachmentName:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/OrderStatus"
        createdAt:
          type: string
          format: date-time
    OrderDetail:
      type: object
      required: [order, history, quotations, dispatchDocuments, payments]
      properties:
        order:
          $ref: "#/components/schemas/Order"
        history:
          type: array
          items:
            $ref: "#/components/schemas/OrderStatusEvent"
        quotations:
          type: array
          items:
            $ref: "#/components/schemas/Quotation"
        dispatchDocuments:
          type: array
          items:
            $ref: "#/components/schemas/DispatchDocument"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentDetail"
    OrderStatusEvent:
      type: object
      required: [id, orderId, status, updatedByName, createdAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        remarks:
          type: string
          nullable: true
        updatedByName:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Quotation:
      type: object
      required:
        - id
        - orderId
        - revisionNo
        - rate
        - gstPercent
        - freight
        - discount
        - paymentTerms
        - status
        - totalAmount
        - createdAt
      properties:
        id:
          type: string
        orderId:
          type: string
        revisionNo:
          type: integer
        rate:
          type: number
        gstPercent:
          type: number
        freight:
          type: number
        discount:
          type: number
        paymentTerms:
          type: string
        expectedDispatchDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/QuotationStatus"
        responseReason:
          type: string
          nullable: true
        totalAmount:
          type: number
        createdByName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateQuotationRequest:
      type: object
      required: [rate, gstPercent, freight, discount, paymentTerms]
      properties:
        rate:
          type: number
        gstPercent:
          type: number
        freight:
          type: number
        discount:
          type: number
        paymentTerms:
          type: string
        expectedDispatchDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
    QuotationResponseRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, reject, request_revision]
        reason:
          type: string
          nullable: true
    DispatchDocument:
      type: object
      required: [id, orderId, docType, fileName, dispatchDate, createdAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        docType:
          $ref: "#/components/schemas/DispatchDocType"
        fileName:
          type: string
        fileUrl:
          type: string
          nullable: true
        transportDetails:
          type: string
          nullable: true
        vehicleNumber:
          type: string
          nullable: true
        dispatchDate:
          type: string
          format: date
        quantityDispatched:
          type: integer
          nullable: true
        uploadedByName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    Payment:
      type: object
      required:
        - id
        - orderId
        - invoiceAmount
        - paymentReceived
        - balanceOutstanding
        - status
        - createdAt
      properties:
        id:
          type: string
        orderId:
          type: string
        invoiceNumber:
          type: string
          nullable: true
        invoiceAmount:
          type: number
        paymentReceived:
          type: number
        balanceOutstanding:
          type: number
        dueDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: "#/components/schemas/PaymentStatus"
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    PaymentDetail:
      type: object
      required: [payment, proofs]
      properties:
        payment:
          $ref: "#/components/schemas/Payment"
        proofs:
          type: array
          items:
            $ref: "#/components/schemas/PaymentProof"
    CreatePaymentRequest:
      type: object
      required: [invoiceAmount, paymentReceived]
      properties:
        invoiceNumber:
          type: string
          nullable: true
        invoiceAmount:
          type: number
        paymentReceived:
          type: number
        dueDate:
          type: string
          format: date
          nullable: true
        remarks:
          type: string
          nullable: true
    PaymentProof:
      type: object
      required: [id, paymentId, fileName, status, createdAt]
      properties:
        id:
          type: string
        paymentId:
          type: string
        fileName:
          type: string
        fileUrl:
          type: string
          nullable: true
        amountClaimed:
          type: number
          nullable: true
        payerRemarks:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/PaymentProofStatus"
        reviewRemarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    SubmitPaymentProofRequest:
      type: object
      required: [fileName]
      properties:
        fileName:
          type: string
        fileUrl:
          type: string
          nullable: true
        amountClaimed:
          type: number
          nullable: true
        payerRemarks:
          type: string
          nullable: true
    ReviewPaymentProofRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [approve, reject]
        reviewRemarks:
          type: string
          nullable: true
...